Debounce input to create new study
